refactor: review by @coderabbitai
