+++ packages/editor/src/components/MarkdownPreview.tsx
@@ -18,6 +18,7 @@ import { Statistics } from '@/components/itinerary/Statistics';
 import { createRenderBlock } from '@/components/render';
 import type { MdNode } from '@/components/render/types';
 import { Tags } from '@/components/Tags';
+import { findBestTargetForLine, isElementVisibleWithin, scrollElementIntoCenteredView } from '@/utils/dom';
 
 interface MarkdownPreviewProps {
     content: string;
@@ -57,6 +58,8 @@ const inlineToSegments = (inline?: PhrasingContent[] | null): TextSegment[] | un
 };
 
 const segmentsToPlainText = (segments?: TextSegment[]): string | undefined => (Array.isArray(segments) ? segments.map((s) => s.text).join('') : undefined);
+
+// moved to utils/dom.ts
 
 const MarkdownPreviewComponent: FC<MarkdownPreviewProps> = ({ content, timezone, currency, rate, showPast, title, description, tags, activeLine, autoScroll = true, onShowPast, preferAltNames, externalContainerRef }) => {
     const displayTimezone = timezone || Intl.DateTimeFormat().resolvedOptions().timeZone;
@@ -265,55 +268,10 @@ const MarkdownPreviewComponent: FC<MarkdownPreviewProps> = ({ content, timezone,
             el.classList.remove('itmd-active');
         }
         if (!activeLine) return;
-        // find best matching element using the same heuristic as auto-scroll
-        let nodes = Array.from(container.querySelectorAll<HTMLElement>('[data-itin-line-start], [data-itin-line-end]'));
-        if (nodes.length === 0) nodes = Array.from(container.querySelectorAll<HTMLElement>('[data-sourcepos]'));
-        if (nodes.length === 0) return;
         const line = activeLine;
-        let best: { el: HTMLElement; score: number } | null = null;
-        for (const el of nodes) {
-            const ds = (el as any).dataset || {};
-            let start = Number((ds as any).itinLineStart || (ds as any).lineStart || NaN);
-            let end = Number((ds as any).itinLineEnd || (ds as any).lineEnd || NaN);
-            if ((!Number.isFinite(start) || !Number.isFinite(end)) && typeof (ds as any).sourcepos === 'string') {
-                const parts = ((ds as any).sourcepos as string).split('-');
-                if (parts.length === 2) {
-                    const s = Number(parts[0].split(':')[0]);
-                    const e = Number(parts[1].split(':')[0]);
-                    if (Number.isFinite(s)) start = s;
-                    if (Number.isFinite(e)) end = e;
-                }
-            }
-            const hasStart = Number.isFinite(start);
-            const hasEnd = Number.isFinite(end);
-            let contains = false;
-            if (hasStart && hasEnd) contains = start <= line && line <= end;
-            else if (hasStart) contains = start === line;
-            else if (hasEnd) contains = end === line;
-            if (contains) {
-                best = { el, score: 0 };
-                break;
-            }
-            if (hasStart && start <= line) {
-                const score = line - start;
-                if (!best || score < best.score) best = { el, score };
-            }
-        }
-        const target = best?.el || nodes[0];
+        const { target, boxTarget } = findBestTargetForLine(container, line);
         if (!target) return;
-        // choose a visible box (avoid display: contents wrappers)
-        const hasBox = (el: Element) => el.getClientRects().length > 0;
-        let boxTarget: HTMLElement | null = hasBox(target) ? (target as HTMLElement) : null;
-        if (!boxTarget) {
-            const descendants = target.querySelectorAll<HTMLElement>(' *');
-            for (const el of Array.from(descendants)) {
-                if (hasBox(el)) {
-                    boxTarget = el;
-                    break;
-                }
-            }
-        }
-        const applyEl = boxTarget || (target as HTMLElement);
+        const applyEl = (boxTarget || target) as HTMLElement;
         applyEl.classList.add('itmd-active');
         // auto-remove highlight after 2 seconds
         highlightTimeoutRef.current = window.setTimeout(() => {
@@ -332,65 +290,12 @@ const MarkdownPreviewComponent: FC<MarkdownPreviewProps> = ({ content, timezone,
         if (!autoScroll) return;
         if (!activeLine || !containerRef.current) return;
         const container = containerRef.current;
-        let nodes = Array.from(container.querySelectorAll<HTMLElement>('[data-itin-line-start], [data-itin-line-end]'));
-        if (nodes.length === 0) {
-            nodes = Array.from(container.querySelectorAll<HTMLElement>('[data-sourcepos]'));
-        }
-        if (nodes.length === 0) return;
-        const line = activeLine;
-        let best: { el: HTMLElement; score: number } | null = null;
-        for (const el of nodes) {
-            const ds = el.dataset || {};
-            let start = Number(ds.itinLineStart || ds.lineStart || NaN);
-            let end = Number(ds.itinLineEnd || ds.lineEnd || NaN);
-            if ((!Number.isFinite(start) || !Number.isFinite(end)) && typeof ds.sourcepos === 'string') {
-                const sp = ds.sourcepos;
-                const parts = sp.split('-');
-                if (parts.length === 2) {
-                    const s = parts[0].split(':')[0];
-                    const e = parts[1].split(':')[0];
-                    const sNum = Number(s);
-                    const eNum = Number(e);
-                    if (Number.isFinite(sNum)) start = sNum;
-                    if (Number.isFinite(eNum)) end = eNum;
-                }
-            }
-            const hasStart = Number.isFinite(start);
-            const hasEnd = Number.isFinite(end);
-            let contains = false;
-            if (hasStart && hasEnd) contains = start <= line && line <= end;
-            else if (hasStart) contains = start === line;
-            else if (hasEnd) contains = end === line;
-            if (contains) {
-                best = { el, score: 0 };
-                break;
-            }
-            if (hasStart && start <= line) {
-                const score = line - start;
-                if (!best || score < best.score) best = { el, score };
-            }
-        }
-        const target = best?.el || nodes[0];
+        const { target, boxTarget } = findBestTargetForLine(container, activeLine);
         if (!target) return;
-        const hasBox = (el: Element) => el.getClientRects().length > 0;
-        let boxTarget: HTMLElement | null = hasBox(target) ? target : null;
-        if (!boxTarget) {
-            const descendants = target.querySelectorAll<HTMLElement>(' *');
-            for (const el of Array.from(descendants)) {
-                if (hasBox(el)) {
-                    boxTarget = el;
-                    break;
-                }
-            }
-        }
-        if (!boxTarget) boxTarget = target;
-        const cRect = container.getBoundingClientRect();
-        const tRect = boxTarget.getBoundingClientRect();
-        const margin = 8;
-        const visible = tRect.bottom > cRect.top + margin && tRect.top < cRect.bottom - margin;
+        const applyEl = (boxTarget || target) as HTMLElement;
+        const visible = isElementVisibleWithin(container, applyEl, 8);
         if (visible) return;
-        const delta = tRect.top - cRect.top - container.clientHeight / 2 + tRect.height / 2;
-        container.scrollBy({ top: delta, behavior: 'smooth' });
+        scrollElementIntoCenteredView(container, applyEl);
     }, [activeLine, autoScroll]);
 
     return (
